Crate gpgme

Reexports

pub use self::context::Context;

pub use self::data::Data;

Modules

context

data

edit

engine

error

info
Constants for use with Token::get_dir_info.

keys

notation

ops

trust

Structs

Error
A type wrapping errors produced by GPG libraries.

Protocol
A cryptographic protocol that may be used with the library.

Token
A type for managing global resources within the library.

Validity

Enums

StrError

Constants

PROTOCOL_ASSUAN

PROTOCOL_CMS

PROTOCOL_DEFAULT

PROTOCOL_G13

PROTOCOL_GPGCONF

PROTOCOL_OPENPGP

PROTOCOL_SPAWN

PROTOCOL_UISERVER

PROTOCOL_UNKNOWN

VALIDITY_FULL

VALIDITY_MARGINAL

VALIDITY_NEVER

VALIDITY_ULTIMATE

VALIDITY_UNDEFINED

VALIDITY_UNKNOWN

Traits

Wrapper

Functions

create_context
Creates a new context for cryptographic operations.

init
Initializes the gpgme library.

Type Definitions

Result

StrResult