Struct gpgme::data::Data

pub struct Data<'a> {
    // some fields omitted
}

Methods

impl<'a> Data<'a>

fn stdin() -> Result<Data<'static>>

fn stdout() -> Result<Data<'static>>

fn stderr() -> Result<Data<'static>>

fn new() -> Result<Data<'static>>

Constructs an empty data object.

fn load<P: AsRef<Path> + ?Sized>(path: &P) -> Result<Data<'static>>

Constructs a data object and fills it with the contents of the file referenced by path.

fn from_bytes<B: AsRef<[u8]> + ?Sized>(bytes: &B) -> Result<Data<'static>>

Constructs a data object and fills it with a copy of bytes.

fn from_buffer<B: AsRef<[u8]> + ?Sized>(buf: &B) -> Result<Data>

Constructs a data object which copies from buf as needed.

fn from_fd<T: AsRawFd + ?Sized>(file: &T) -> Result<Data>

unsafe fn from_raw_file<'b>(file: *mut FILE) -> Result<Data<'b>>

fn from_reader<R: Send + 'static>(r: R) -> Result<Data<'static>, WrappedError<R>> where R: Read

fn from_seekable_reader<R: Send + 'static>(r: R) -> Result<Data<'static>, WrappedError<R>> where R: Read + Seek

fn from_writer<W: Send + 'static>(w: W) -> Result<Data<'static>, WrappedError<W>> where W: Write

fn from_seekable_writer<W: Send + 'static>(w: W) -> Result<Data<'static>, WrappedError<W>> where W: Write + Seek

fn from_stream<S: Send + 'static>(s: S) -> Result<Data<'static>, WrappedError<S>> where S: Read + Write

fn from_seekable_stream<S: Send + 'static>(s: S) -> Result<Data<'static>, WrappedError<S>> where S: Read + Write + Seek

fn filename(&self) -> Option<&str>

fn clear_filename(&mut self) -> Result<()>

fn set_filename<S: Into<String>>(&mut self, name: S) -> Result<()>

fn encoding(&self) -> Encoding

fn set_encoding(&mut self, enc: Encoding) -> Result<()>

fn identify(&mut self) -> Type

fn into_bytes(self) -> Option<Vec<u8>>

fn into_string(self) -> Result<String, Option<FromUtf8Error>>