Struct gpgme::Token

pub struct Token(_);

A type for managing global resources within the library.

Methods

impl Token

fn check_version<S: Into<String>>(&self, version: S) -> bool

Checks that the linked version of the library is at least the specified version.

Note: false is returned, if version is not in the format MAJOR.MINOR.MICRO.

Examples

let gpgme = gpgme::init();
assert!(gpgme.check_version("1.4.0"));

fn version(&self) -> &'static str

Returns the version string for the library.

fn get_dir_info<S: Into<String>>(&self, what: S) -> Option<&'static str>

Returns the default value for specified configuration option.

Commonly supported values for what are specified in info.

This function requires a version of GPGme >= 1.5.0.

fn check_engine_version(&self, proto: Protocol) -> Result<()>

Checks that the engine implementing the protocol proto meets requirements of the library.

fn engine_info(&self) -> Result<EngineInfoGuard>

fn set_engine_info(&self, proto: Protocol, filename: Option<String>, home_dir: Option<String>) -> Result<()>